+++ bin/integration-tests/src/tests/client.rs
@@ -373,9 +373,6 @@ pub async fn test_import_expected_notes_from_the_past_as_committed(
 
     execute_tx_and_sync(&mut client_1, faucet_account.id(), tx_request).await?;
 
-    // Use client 1 to wait until a couple of blocks have passed
-    wait_for_blocks(&mut client_1, 3).await;
-
     // importing the note before client_2 is synced will result in a note with `Expected` state
     let note_id = client_2
         .import_notes(&[NoteFile::NoteDetails {