+++ crates/web-client/js/workers/web-client-methods-worker.js
@@ -348,8 +348,14 @@ async function processMessage(event) {
         hasGetKeyCb,
         hasInsertKeyCb,
         hasSignCb,
+        logLevel,
       ] = args;
       const wasm = await getWasmOrThrow();
+
+      if (logLevel) {
+        wasm.setupLogging(logLevel);
+      }
+
       wasmWebClient = new wasm.WebClient();
 
       // Check if any callbacks are provided
@@ -380,8 +386,13 @@ async function processMessage(event) {
       self.postMessage({ ready: true });
       return;
     } else if (action === WorkerAction.INIT_MOCK) {
-      const [seed] = args;
+      const [seed, logLevel] = args;
       const wasm = await getWasmOrThrow();
+
+      if (logLevel) {
+        wasm.setupLogging(logLevel);
+      }
+
       wasmWebClient = new wasm.WebClient();
       await wasmWebClient.createMockClient(seed);